client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.0",
    "mcp>=1.9.4",
    "python-dotenv>=1.1.0",
]
